Update do Pedido ainda não esta completo

+++ aplicacao/uPedidoCadastroRegra.pas
@@ -130,13 +130,16 @@ begin
   Result := 0;
   if (aPedidoDTO.idPedido > 0) then
   begin
+    if (aModel.DeleteItemPedidoSabores(aPedidoDTO.idPedido)) then
+    begin
      if (aModel.DeleteItensPedido(aPedidoDTO.idPedido)) then
      begin
        if (not(aModel.InsertItensPedido(aPedidoDTO))) then
          result := 1;
      end
      else
-      Result := 2
+      Result := 2;
+    end
   end
   else
   begin
